fixed Ivan Yarymovych's comments
