--- src/app/body/all-products/all-products/all-products.component.ts
@@ -1,23 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { ProductService } from '../../../services/product.service';
-import { Product } from '../../../models/product.model';
-
-@Component({
-  selector: 'app-all-products',
-  templateUrl: './all-products.component.html',
-  styleUrls: ['./all-products.component.scss']
-})
-export class AllProductsComponent implements OnInit {
-  public products: Product[];
-  constructor(private productService: ProductService) { }
-  ngOnInit(): void {
-    this.getProducts();
-  }
-
-  getProducts(): void {
-    this.productService.getProducts()
-      .subscribe(products => this.products = products);
-  }
-}
-
-
