+++ src/app/body/all-products/filters/rating-filter/rating-filter.component.html
@@ -1,10 +1,10 @@
 <div class="filter_brand">
   <label>
-    <img src="{{starFull}}" alt="">
-    <img src="{{starFull}}" alt="">
-    <img src="{{starFull}}" alt="">
-    <img src="{{starFull}}" alt="">
-    <img src="{{starFull}}" alt="">
+    <img src="assets/star-full.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
     <input type="checkbox" />
     <span></span>
   </label>
@@ -12,11 +12,11 @@
 
 <div class="filter_brand">
   <label>
-    <img src="{{star}}" alt="">
-    <img src="{{starFull}}" alt="">
-    <img src="{{starFull}}" alt="">
-    <img src="{{starFull}}" alt="">
-    <img src="{{starFull}}" alt="">
+    <img src="assets/star.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
     <input type="checkbox" />
     <span></span>
   </label>
@@ -24,11 +24,11 @@
 
 <div class="filter_brand">
   <label>
-    <img src="{{star}}" alt="">
-    <img src="{{star}}" alt="">
-    <img src="{{starFull}}" alt="">
-    <img src="{{starFull}}" alt="">
-    <img src="{{starFull}}" alt="">
+    <img src="assets/star.svg" alt="">
+    <img src="assets/star.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
     <input type="checkbox" />
     <span></span>
   </label>
@@ -36,11 +36,11 @@
 
 <div class="filter_brand">
   <label>
-    <img src="{{star}}" alt="">
-    <img src="{{star}}" alt="">
-    <img src="{{star}}" alt="">
-    <img src="{{starFull}}" alt="">
-    <img src="{{starFull}}" alt="">
+    <img src="assets/star.svg" alt="">
+    <img src="assets/star.svg" alt="">
+    <img src="assets/star.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
     <input type="checkbox" />
     <span></span>
   </label>
@@ -48,11 +48,11 @@
 
 <div class="filter_brand">
   <label>
-    <img src="{{star}}" alt="">
-    <img src="{{star}}" alt="">
-    <img src="{{star}}" alt="">
-    <img src="{{star}}" alt="">
-    <img src="{{starFull}}" alt="">
+    <img src="assets/star.svg" alt="">
+    <img src="assets/star.svg" alt="">
+    <img src="assets/star.svg" alt="">
+    <img src="assets/star.svg" alt="">
+    <img src="assets/star-full.svg" alt="">
     <input type="checkbox" />
     <span></span>
   </label>
