+++ src/app/body/all-products/filters/rating-filter/rating-filter.component.ts
@@ -1,8 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-// @ts-ignore
-import star from '../../../../../assets/star.svg';
-// @ts-ignore
-import starFull from '../../../../../assets/star-full.svg';
 
 @Component({
   selector: 'app-rating-filter',
@@ -10,14 +6,11 @@ import starFull from '../../../../../assets/star-full.svg';
   styleUrls: ['./rating-filter.component.scss']
 })
 export class RatingFilterComponent implements OnInit {
-  public star;
-  public starFull;
 
   constructor() { }
 
   ngOnInit(): void {
-    this.star = star;
-    this.starFull = starFull;
+
   }
 
 }
