+++ src/app/body/all-products/product-card/product-card.component.html
@@ -8,11 +8,7 @@
       <div class="product-card_title">{{product.title}}</div>
       <div class="product-card_description openSans">{{product.description.small}}</div>
       <div class="product-card_rating">
-        <img src="{{starBlack}}" alt="">
-        <img src="{{product.rating > 1 ? starBlack : star}}" alt="">
-        <img src="{{product.rating > 2 ? starBlack : star}}" alt="">
-        <img src="{{product.rating > 3 ? starBlack : star}}" alt="">
-        <img src="{{product.rating > 4 ? starBlack : star}}" alt="">
+        <img *ngFor="let star of productRating" [src]="'assets/' + star + '.svg'" alt="star">
       </div>
       <div class="product-card_info openSans">
         <div>
@@ -43,12 +39,7 @@
         <a routerLink="/products/{{product.id}}">Product Detail</a>
       </div>
       <div class="product-card_wishlist">
-        <button>
-          <svg width="17" height="14" viewBox="0 0 17 14" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path fill-rule="evenodd" clip-rule="evenodd" d="M9.13062 2.26001C9.76714 1.62349 10.6304 1.2659 11.5306 1.2659C12.4308 1.2659 13.2941 1.62349 13.9306 2.26001C14.5671 2.89653 14.9247 3.75984 14.9247 4.66001C14.9247 5.56019 14.5671 6.42349 13.9306 7.06001L13.0573 7.93334L8.25729 12.7333L3.45729 7.93334L2.58396 7.06001C1.94744 6.42349 1.58984 5.56019 1.58984 4.66001C1.58984 3.75984 1.94744 2.89653 2.58396 2.26001C3.22048 1.62349 4.08378 1.2659 4.98396 1.2659C5.88413 1.2659 6.74744 1.62349 7.38396 2.26001L8.25729 3.13334L9.13062 2.26001Z" stroke="#151515" stroke-width="2.5" stroke-linecap="round" stroke-linejoin="round"/>
-          </svg>
-          Add to wish list
-        </button>
+        <button>Add to wish list</button>
       </div>
     </div>
   </div>
