+++ src/app/body/product/product.component.html
@@ -1,9 +1,9 @@
 <div class="product_wrapper" *ngIf="product">
-  <div class="product_fotos">
-    <div class="product_fotos-top">
+  <div class="product_photos">
+    <div class="product_photos-top">
       <img src="assets/imageProdMain.jpg" alt="">
     </div>
-    <div class="product_fotos-bottom">
+    <div class="product_photos-bottom">
       <div><img src="assets/imageProdMain.jpg" alt=""></div>
       <div><img src="assets/imageProdMain.jpg" alt=""></div>
     </div>
@@ -12,11 +12,7 @@
   <div class="product_content">
     <div class="product_title">{{product.title.charAt(0).toUpperCase() + product.title.slice(1)}}</div>
     <div class="product_rating">
-      <img src="{{starBlack}}" alt="">
-      <img src="{{product.rating > 1 ? starBlack : star}}" alt="">
-      <img src="{{product.rating > 2 ? starBlack : star}}" alt="">
-      <img src="{{product.rating > 3 ? starBlack : star}}" alt="">
-      <img src="{{product.rating > 4 ? starBlack : star}}" alt="">
+      <img *ngFor="let star of productRating" [src]="'assets/' + star + '.svg'" alt="star">
       <span class="openSans">({{product.reviews.length}} customer review{{product.reviews.length > 1 ? 's' : ''}})</span>
     </div>
     <div class="product_description openSans">{{product.description.main}}</div>
@@ -32,7 +28,7 @@
         </div>
         <div>
           <span>Stock:</span>
-          <span>{{product.stock}}</span>
+          <span>{{product.stock >= 10 ? 'In Stock' : product.stock === 0 ? 'Ran out' : 'Less than 10 items'}}</span>
         </div>
         <div>
           <span>Color:</span>
@@ -42,7 +38,7 @@
       <div class="product_info-part">
         <div>
           <span>Size:</span>
-          <span>{{product.country}}</span>
+          <span>{{product.sizes}}</span>
         </div>
         <div>
           <span>Buy by:</span>
@@ -74,21 +70,12 @@
       </div>
 
       <div class="product_button">
-        <button>
-          <svg width="14" height="14" viewBox="0 0 14 14" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M12.1608 7H1.49414" stroke="#92C064" stroke-width="2.5" stroke-linecap="round" stroke-linejoin="bevel"/>
-            <path d="M6.82812 12.3333V1.66663" stroke="#92C064" stroke-width="2.5" stroke-linecap="round" stroke-linejoin="bevel"/>
-          </svg>
-          Add to cart</button>
+        <button>Add to cart</button>
       </div>
     </div>
 
     <div class="product_wishlist">
-      <button>
-        <svg width="17" height="14" viewBox="0 0 17 14" fill="none" xmlns="http://www.w3.org/2000/svg">
-          <path fill-rule="evenodd" clip-rule="evenodd" d="M9.13062 2.26001C9.76714 1.62349 10.6304 1.2659 11.5306 1.2659C12.4308 1.2659 13.2941 1.62349 13.9306 2.26001C14.5671 2.89653 14.9247 3.75984 14.9247 4.66001C14.9247 5.56019 14.5671 6.42349 13.9306 7.06001L13.0573 7.93334L8.25729 12.7333L3.45729 7.93334L2.58396 7.06001C1.94744 6.42349 1.58984 5.56019 1.58984 4.66001C1.58984 3.75984 1.94744 2.89653 2.58396 2.26001C3.22048 1.62349 4.08378 1.2659 4.98396 1.2659C5.88413 1.2659 6.74744 1.62349 7.38396 2.26001L8.25729 3.13334L9.13062 2.26001Z" stroke="#151515" stroke-width="2.5" stroke-linecap="round" stroke-linejoin="round"/>
-        </svg>
-        Add to my wish list</button>
+      <button>Add to my wish list</button>
     </div>
   </div>
