+++ src/app/body/product/product.component.scss
@@ -3,7 +3,7 @@
     display: flex;
     justify-content: space-between;
   }
-  &_fotos {
+  &_photos {
     display: flex;
     flex-direction: column;
     margin-right: 32px;
@@ -116,16 +116,10 @@
     }
   }
   &_button {
-    svg {
-      margin-right: 5px;
-      path {
-        transition: .3s;
-      }
-    }
     button {
-      background: #6A983C;
+      background: url("/assets/plus.svg") 10% center no-repeat, #6A983C;
       border: 2px solid #46760A;
-      padding: 12px 16px;
+      padding: 12px 16px 12px 30px;
       color: #FFFFFF;
       font-weight: bold;
       font-size: 15px;
@@ -135,32 +129,24 @@
       display: flex;
       align-items: center;
       &:hover {
-        background: #46760A;
-        path {
-          stroke: #fff;
-        }
+        background: url("/assets/plus.svg") 10% center no-repeat, #46760A;
       }
     }
   }
   &_wishlist {
     margin-top: 24px;
-    padding-left: 10px;
     button {
       font-weight: bold;
       font-size: 15px;
       border: none;
-      background: none;
       cursor: pointer;
-      svg {
-        margin-right: 5px;
-        path {
-          transition: .3s;
-        }
-      }
+      padding: 12px 15px 12px 40px;
+      transition: .3s;
+      background: url("/assets/heart.svg") 10% center no-repeat;
+      border-radius: 12px;
       &:hover {
-        svg path {
-          fill: #000;
-        }
+        background: url("/assets/heart-full.svg") 10% center no-repeat;
+        background-color: #F5F5F5;
       }
     }
   }
