+++ src/app/body/product/product.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { ProductRoutingModule } from './product-routing.module';
-import { ProductComponent } from './product/product.component';
+import { ProductComponent } from './product.component';
 
 
 @NgModule({
