+++ src/app/resolvers/product.resolver.ts
@@ -6,10 +6,10 @@ import { Product } from '../models/product.model';
 
 
 @Injectable()
-export class ProductsResolver implements Resolve<Observable<Product[]>> {
+export class ProductResolver implements Resolve<Observable<Product>> {
   constructor(private productService: ProductService) {}
 
   resolve(route: ActivatedRouteSnapshot): Observable<any> {
-    return of(this.productService.getProducts());
+    return of(this.productService.getProductById(route.params.id));
   }
 }
